Fix mgmt route VRF name on cat9kv
VRF for the static default management route was incorrectly set to `clab-mgmt` which is the name for the management VRF on other IOS-XE devices.

Cat9kv automatically creates it's own management VRF with the 'Mgmt-vrf' name. I most likely made an erroneous copy and paste. 

Simple fix to correct this error. Apologies.